refactor(WebAPI): Moved GetAllMediaEpisodesCommand to vertical slice
