+++ src/Application/PlexLibraries/RefreshLibraryMedia/RefreshLibraryMediaCommand.cs
@@ -103,118 +103,12 @@ public class RefreshLibraryMediaCommandHandler : IRequestHandler<RefreshLibraryM
             case PlexMediaType.Movie:
                 return await RefreshPlexMovieLibrary(newPlexLibrary);
             case PlexMediaType.TvShow:
-                return await RefreshPlexTvShowLibrary(newPlexLibrary);
+                return await _mediator.Send(new RefreshPlexTvShowLibraryCommand(newPlexLibrary), cancellationToken);
             default:
                 return Result
                     .Fail($"Library type {newPlexLibrary.Type} is currently not supported by PlexRipper")
                     .LogWarning();
         }
-    }
-
-    private async Task<Result<PlexLibrary>> RefreshPlexTvShowLibrary(PlexLibrary plexLibrary)
-    {
-        if (plexLibrary.Type != PlexMediaType.TvShow)
-            return Result.Fail("PlexLibrary is not of type TvShow").LogError();
-
-        if (plexLibrary.TvShows.Any())
-        {
-            var timer = new Stopwatch();
-            timer.Start();
-
-            // Phase 2 of 5: Season data was retrieved successfully.
-            var rawSeasonDataResult = await _plexServiceApi.GetAllSeasonsAsync(
-                plexLibrary,
-                progress => SendProgress(2, progress.Percentage, progress.TimeRemaining)
-            );
-
-            if (rawSeasonDataResult.IsFailed)
-                return rawSeasonDataResult.ToResult();
-
-            // Phase 3 of 5: Episode data was retrieved successfully.
-            var rawEpisodesDataResult = await _plexServiceApi.GetAllEpisodesAsync(
-                plexLibrary,
-                progress => SendProgress(3, progress.Percentage, progress.TimeRemaining)
-            );
-            if (rawEpisodesDataResult.IsFailed)
-                return rawEpisodesDataResult.ToResult();
-
-            _log.Information("Merging all data received from PlexApi for library {PlexLibraryName}", plexLibrary.Name);
-
-            // Phase 4 of 5: PlexLibrary media data was parsed successfully.
-            _log.Here()
-                .Debug(
-                    "Finished retrieving all media for library {PlexLibraryName} in {Elapsed:000} seconds",
-                    plexLibrary.Title,
-                    timer.Elapsed.TotalSeconds
-                );
-            timer.Restart();
-
-            var rawSeasonData = rawSeasonDataResult.Value;
-            var rawEpisodesData = rawEpisodesDataResult.Value;
-
-            // Phase 4 of 5: PlexLibrary media data was parsed successfully.
-            var tvShows = BuildTvShowTree(plexLibrary, plexLibrary.TvShows, rawSeasonData, rawEpisodesData);
-            SendProgress(4, 1);
-
-            // Update the MetaData of this library
-            var syncResult = await _mediator.Send(new SyncPlexTvShowsCommand(tvShows));
-            if (syncResult.IsFailed)
-            {
-                SendProgress(5, 1);
-                return syncResult.ToResult().LogError();
-            }
-
-            var mediaSize = tvShows.Sum(x => x.MediaSize);
-            plexLibrary.SetTvShowMetaData(
-                plexLibrary.TvShows.Count,
-                rawSeasonData.Count,
-                rawEpisodesData.Count,
-                mediaSize
-            );
-
-            if (plexLibrary.TvShows.Any() && mediaSize == 0)
-            {
-                _log.Error(
-                    "No media size was found for library {PlexLibraryName} with id: {PlexLibraryId}",
-                    plexLibrary.Title,
-                    plexLibrary.Id
-                );
-            }
-
-            await _dbContext.UpdatePlexLibraryById(plexLibrary);
-
-            _log.Here()
-                .Debug(
-                    "Finished updating all media in the database for library {PlexLibraryName} in {Elapsed:0} seconds",
-                    plexLibrary.Title,
-                    timer.Elapsed.TotalSeconds
-                );
-
-            // Phase 5 of 5: Database has been successfully updated with new library data.
-            SendProgress(5, 1);
-        }
-        else
-        {
-            _log.Warning(
-                "No TV shows were found for library {PlexLibraryName} with id: {PlexLibraryId}",
-                plexLibrary.Title,
-                plexLibrary.Id
-            );
-        }
-
-        // Mark the library as synced
-        plexLibrary.SyncedAt = DateTime.UtcNow;
-        await _dbContext
-            .PlexLibraries.Where(x => x.Id == plexLibrary.Id)
-            .ExecuteUpdateAsync(p => p.SetProperty(x => x.SyncedAt, plexLibrary.SyncedAt));
-
-        _log.Information(
-            "Successfully refreshed library {PlexLibraryName} with id: {PlexLibraryId}",
-            plexLibrary.Title,
-            plexLibrary.Id
-        );
-
-        return Result.Ok(plexLibrary);
     }
 
     private async Task<Result<PlexLibrary>> RefreshPlexMovieLibrary(PlexLibrary plexLibrary)
@@ -295,138 +189,5 @@ public class RefreshLibraryMediaCommandHandler : IRequestHandler<RefreshLibraryM
         _progressAction?.Invoke(progress);
 
         _signalRService.SendLibraryProgressUpdateAsync(progress);
-    }
-
-    private List<PlexTvShow> BuildTvShowTree(
-        PlexLibrary plexLibrary,
-        List<PlexTvShow> rawTvShowData,
-        List<PlexTvShowSeason> rawSeasonData,
-        List<PlexTvShowEpisode> rawEpisodesData
-    )
-    {
-        var (validSeasons, validEpisodes) = Filter(rawSeasonData, rawEpisodesData, plexLibrary);
-
-        // Group seasons and episodes by parent key upfront
-        var seasonsByTvShowKey = validSeasons.GroupBy(x => x.ParentGuid!).ToDictionary(g => g.Key, g => g.ToList());
-        var episodesBySeasonKey = validEpisodes.GroupBy(x => x.ParentGuid!).ToDictionary(g => g.Key, g => g.ToList());
-
-        for (var i = 0; i < rawTvShowData.Count; i++)
-        {
-            var plexTvShow = rawTvShowData[i];
-            plexTvShow.PlexLibraryId = plexLibrary.Id;
-            plexTvShow.PlexServerId = plexLibrary.PlexServerId;
-            plexTvShow.SortIndex = i + 1;
-
-            // Retrieve and assign seasons for this TV show
-            if (seasonsByTvShowKey.TryGetValue(plexTvShow.Guid, out var seasons))
-            {
-                plexTvShow.Seasons = seasons;
-                plexTvShow.ChildCount = seasons.Count;
-
-                // Remove seasons that have been assigned
-                seasonsByTvShowKey.Remove(plexTvShow.Guid);
-            }
-
-            for (var seasonIndex = 0; seasonIndex < plexTvShow.Seasons.Count; seasonIndex++)
-            {
-                var plexTvShowSeason = plexTvShow.Seasons[seasonIndex];
-                plexTvShowSeason.SortIndex = seasonIndex + 1;
-                plexTvShowSeason.PlexLibraryId = plexLibrary.Id;
-                plexTvShowSeason.PlexServerId = plexLibrary.PlexServerId;
-                plexTvShowSeason.TvShow = plexTvShow;
-
-                // Retrieve and assign episodes for this season
-                if (!episodesBySeasonKey.TryGetValue(plexTvShowSeason.Guid, out var episodes))
-                    continue;
-
-                // Set library ID in each episode
-                var episodeIndex = 1;
-                episodes.ForEach(
-                    (x) =>
-                    {
-                        x.PlexLibraryId = plexLibrary.Id;
-                        x.PlexServerId = plexLibrary.PlexServerId;
-                        x.SortIndex = episodeIndex++;
-                    }
-                );
-
-                plexTvShowSeason.Episodes = episodes;
-                plexTvShowSeason.ChildCount = episodes.Count;
-
-                // Remove episodes that have been assigned
-                episodesBySeasonKey.Remove(plexTvShowSeason.Guid);
-
-                // Set the season's year based on the first episode's year
-                if (plexTvShowSeason.Year == 0 && episodes.Any())
-                    plexTvShowSeason.Year = episodes.First().Year;
-
-                plexTvShowSeason.MediaSize = episodes.Sum(x => x.MediaSize);
-                plexTvShowSeason.Duration = episodes.Sum(x => x.Duration);
-            }
-
-            plexTvShow.MediaSize = plexTvShow.Seasons.Sum(x => x.MediaSize);
-            plexTvShow.Duration = plexTvShow.Seasons.Sum(x => x.Duration);
-            plexTvShow.GrandChildCount = plexTvShow.Seasons.Sum(x => x.ChildCount);
-        }
-
-        return rawTvShowData;
-    }
-
-    private (List<PlexTvShowSeason> validSeasons, List<PlexTvShowEpisode> validEpisodes) Filter(
-        List<PlexTvShowSeason> rawSeasonData,
-        List<PlexTvShowEpisode> rawEpisodesData,
-        PlexLibrary library
-    )
-    {
-        var validSeasons = new List<PlexTvShowSeason>();
-        var inValidSeasons = new List<PlexTvShowSeason>();
-
-        var validEpisodes = new List<PlexTvShowEpisode>();
-        var inValidEpisodes = new List<PlexTvShowEpisode>();
-
-        foreach (var plexTvShowSeason in rawSeasonData)
-        {
-            if (plexTvShowSeason.ParentGuid != null)
-            {
-                validSeasons.Add(plexTvShowSeason);
-                continue;
-            }
-
-            inValidSeasons.Add(plexTvShowSeason);
-        }
-
-        foreach (var episode in rawEpisodesData)
-        {
-            if (episode.ParentGuid != null)
-            {
-                validEpisodes.Add(episode);
-                continue;
-            }
-
-            inValidEpisodes.Add(episode);
-        }
-
-        // Log invalid seasons and episodes
-        if (inValidSeasons.Any())
-        {
-            _log.Warning(
-                "Found {Count} invalid seasons which are missing a ParentGUID in library {PlexLibraryName} with id: {PlexLibraryId}",
-                inValidSeasons.Count,
-                library.Title,
-                library.Id
-            );
-        }
-
-        if (inValidEpisodes.Any())
-        {
-            _log.Warning(
-                "Found {Count} invalid episodes which are missing a ParentGUID in library {PlexLibraryName} with id: {PlexLibraryId}",
-                inValidEpisodes.Count,
-                library.Title,
-                library.Id
-            );
-        }
-
-        return (validSeasons, validEpisodes);
     }
 }
