+++ src/PlexApi.Contracts/PlexMedia/GetLibraryMedia/GetLibraryMediaCommand.cs
@@ -1,0 +1,8 @@
+using FastEndpoints;
+using FluentResults;
+using PlexRipper.Domain;
+
+namespace PlexApi.Contracts.GetLibraryMedia;
+
+public record GetLibraryMediaCommand(PlexLibrary PlexLibrary, Action<MediaSyncProgress>? Action = null)
+    : ICommand<Result<LibraryMetadata>>;
